chore: lib 폴더를 전부 common, modules 이하 폴더로 이동
