+++ src/common/globals.css
