+++ src/common/hooks/use-const.ts
