+++ src/modules/admin-ui/property-panel/TextInputGroup.tsx
@@ -1,5 +1,5 @@
-import useConst from "@lib/hooks/use-const";
-import useEditPropertyInternalValue from "@lib/hooks/use-edit-property-internal-value";
+import useConst from "@common/hooks/use-const";
+import useEditPropertyInternalValue from "@common/hooks/use-edit-property-internal-value";
 import { nanoid } from "nanoid";
 import type { ChangeEvent } from "react";
 import { memo, useCallback } from "react";
