+++ src/modules/editor/components/CodeBlockEditModal.tsx
@@ -1,8 +1,8 @@
 /* This example requires Tailwind CSS v2.0+ */
+import { useHotkeys } from "@common/hooks/use-hotkeys";
+import { useMyStoreMemo } from "@common/store";
 import { convertCodeBlockToString } from "@common/util";
 import { Dialog } from "@headlessui/react";
-import { useHotkeys } from "@lib/hooks/use-hotkeys";
-import { useMyStoreMemo } from "@lib/store";
 import type {
   ICodeBlock,
   ICodeBlockElement,
