+++ src/modules/layout/AdminLayout.tsx
@@ -1,17 +1,17 @@
 import Link from "next/link";
 import type { ReactNode } from "react";
 
-export const adminLayoutWrapper = (page: ReactNode) => {
+export default function adminLayoutWrapper(page: ReactNode) {
   return (
     <div className="flex mx-auto">
-      <div className="p-8 w-64">
+      <div className="w-64 p-8">
         <h2 className="text-lg font-bold">관리자</h2>
         <hr className="mt-3 mb-4" />
         <Link href="/admin/post/list">
           <a>게시글 목록</a>
         </Link>
       </div>
-      <main className="m-8 flex-1">{page}</main>
+      <main className="flex-1 m-8">{page}</main>
     </div>
   );
-};
+}
