+++ src/pages/admin/post/new.tsx
@@ -1,10 +1,10 @@
-import { NoLayoutWrapper } from "@lib/components/layout/NoLayout";
-import PostEditorWrapper from "@lib/components/PostEditorWrapper";
-import { useAdminPageGuard } from "@lib/hooks";
-import useToast from "@lib/hooks/use-toast";
-import { getCreatePostInput } from "@lib/serialize";
-import { editingPostAtom } from "@lib/store";
+import { useAdminPageGuard } from "@common/hooks";
+import useToast from "@common/hooks/use-toast";
+import { editingPostAtom } from "@common/store";
+import { getCreatePostInput } from "@common/util";
 import type { MonnomlogPage } from "@modules/content/types";
+import PostEditorWrapper from "@modules/editor/components/PostEditorWrapper";
+import NoLayoutWrapper from "@modules/layout/NoLayout";
 import type { Post } from "@modules/supabase/supabase";
 import axiosGlobal from "axios";
 import { useAtom } from "jotai";
